Use `InnerHTML` mixin for `innerHTML` definition
