+++ lib/jsdom/living/nodes/Element-impl.js
@@ -10,6 +10,7 @@ const attributes = require("../attributes");
 const namedPropertiesWindow = require("../named-properties-window");
 const NODE_TYPE = require("../node-type");
 const { parseFragment } = require("../../browser/parser");
+const InnerHTMLImpl = require("../domparsing/InnerHTML-impl").implementation;
 const { fragmentSerialization } = require("../domparsing/serialization");
 const { domSymbolTree } = require("../helpers/internal-constants");
 const DOMException = require("domexception/webidl2js-wrapper");
@@ -187,24 +188,6 @@ class ElementImpl extends NodeImpl {
 
     const contextObjectParent = domSymbolTree.parent(this);
     contextObjectParent._replace(fragment, this);
-  }
-
-  // https://w3c.github.io/DOM-Parsing/#dfn-innerhtml
-  get innerHTML() {
-    return fragmentSerialization(this, {
-      requireWellFormed: true,
-      globalObject: this._globalObject
-    });
-  }
-  set innerHTML(markup) {
-    const fragment = parseFragment(markup, this);
-
-    let contextObject = this;
-    if (this.localName === "template" && this.namespaceURI === HTML_NS) {
-      contextObject = contextObject._templateContents;
-    }
-
-    contextObject._replaceAll(fragment);
   }
 
   get classList() {
@@ -566,6 +549,7 @@ mixin(ElementImpl.prototype, NonDocumentTypeChildNode.prototype);
 mixin(ElementImpl.prototype, ParentNodeImpl.prototype);
 mixin(ElementImpl.prototype, ChildNodeImpl.prototype);
 mixin(ElementImpl.prototype, SlotableMixinImpl.prototype);
+mixin(ElementImpl.prototype, InnerHTMLImpl.prototype);
 
 ElementImpl.prototype.getElementsByTagName = memoizeQuery(function (qualifiedName) {
   return listOfElementsWithQualifiedName(qualifiedName, this);
