+++ lib/jsdom/living/nodes/Element.webidl
@@ -51,7 +51,6 @@ dictionary ShadowRootInit {
 
 // https://w3c.github.io/DOM-Parsing/#extensions-to-the-element-interface
 partial interface Element {
-  [CEReactions] attribute [LegacyNullToEmptyString] DOMString innerHTML;
   [CEReactions] attribute [LegacyNullToEmptyString] DOMString outerHTML;
   [CEReactions] void insertAdjacentHTML(DOMString position, DOMString text);
 };
