+++ lib/jsdom/living/nodes/ShadowRoot-impl.js
@@ -1,12 +1,11 @@
 "use strict";
 
-const { parseFragment } = require("../../browser/parser");
-const { fragmentSerialization } = require("../domparsing/serialization.js");
 const { nodeRoot } = require("../helpers/node");
 const { mixin } = require("../../utils");
 
 const DocumentFragment = require("./DocumentFragment-impl").implementation;
 const DocumentOrShadowRootImpl = require("./DocumentOrShadowRoot-impl").implementation;
+const InnerHTMLImpl = require("../domparsing/InnerHTML-impl").implementation;
 
 class ShadowRootImpl extends DocumentFragment {
   constructor(globalObject, args, privateData) {
@@ -31,21 +30,10 @@ class ShadowRootImpl extends DocumentFragment {
   get host() {
     return this._host;
   }
-
-  // https://w3c.github.io/DOM-Parsing/#dfn-innerhtml
-  get innerHTML() {
-    return fragmentSerialization(this, {
-      requireWellFormed: true,
-      globalObject: this._globalObject
-    });
-  }
-  set innerHTML(markup) {
-    const fragment = parseFragment(markup, this._host);
-    this._replaceAll(fragment);
-  }
 }
 
 mixin(ShadowRootImpl.prototype, DocumentOrShadowRootImpl.prototype);
+mixin(ShadowRootImpl.prototype, InnerHTMLImpl.prototype);
 
 module.exports = {
   implementation: ShadowRootImpl
