+++ lib/jsdom/living/nodes/ShadowRoot.webidl
@@ -3,9 +3,6 @@
 interface ShadowRoot : DocumentFragment {
   readonly attribute ShadowRootMode mode;
   readonly attribute Element host;
-
-  // https://github.com/w3c/DOM-Parsing/issues/21
-  [CEReactions] attribute [LegacyNullToEmptyString] DOMString innerHTML;
 };
 
 enum ShadowRootMode { "open", "closed" };
